# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EAMS API",
    "description": "Events Attendance Monitoring System — attendance capture and administration.\n\n**This document supersedes `docs/api/attendance-contract-handoff.md` as the contract.**\nIt is generated from the source, so the outcome tokens, status codes and payload shapes\nbelow are what the running build actually does.\n\n- **Base path** `/api/v1`. JSON in and out, camelCase field names.\n- **All timestamps are UTC**, ISO 8601, with an explicit `Z`. A bare or offset-bearing\n  local time misjudges the Present/Late boundary by the server's offset.\n- **Errors are RFC 7807** `application/problem+json`. Every body carries a `traceId` to\n  quote back and, on the attendance surface, a stable `code` to branch on — never the\n  human-readable `title` or `detail`, which are reworded freely.\n- **Capture endpoints require a device key** (`DeviceKey` below). Everything else is open\n  for now: authentication for human users is Phase 6, and until it lands this API must\n  stay on a local or trusted network.\n\nSee the `TapOutcomeCode`, `ManualOutcomeCode` and `LiveOutcomeCode` schemas for the frozen\ntoken tables.",
    "version": "v1"
  },
  "paths": {
    "/api/v1/academic/terms": {
      "get": {
        "tags": [
          "Academic"
        ],
        "summary": "`GET /academic/terms` — every term, current first and then newest first.",
        "description": "The order is deliberate and does not depend on `startsOn`: the roster source has no term\r\ndate columns at all, so those two fields are frequently null and sorting on them would produce\r\nan arbitrary list on real data. Term codes are operator-authored and sort chronologically by\r\nconstruction, which is what this falls back to.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number, default 1. Out-of-range values are clamped, not refused.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 50, maximum 200; a larger value is clamped and the response says so.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of terms, possibly empty.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TermDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/terms/current": {
      "get": {
        "tags": [
          "Academic"
        ],
        "summary": "`GET /academic/terms/current` — the term flagged current, or 404 when none is.",
        "description": "\r\n\r\n  <b>404 rather than a 200 carrying null, because the two are different facts and a client has to\r\n             act on them differently.</b> \"No term is flagged current\" is a configuration state an\r\n             administrator fixes — it happens between semesters, before anyone has moved the flag — whereas\r\n             an empty 200 body reads as \"here is the current term, it just has no fields\". A caller\r\n             defaulting a term picker needs to tell those apart to know whether to prompt.\r\n             \r\n\r\n\r\n  <b>At most one term can be current, and the database is what guarantees it</b> — a filtered\r\n             unique index, not a convention. Two \"current\" terms would make every term-defaulting query pick\r\n             one at random, which surfaces months later as a report quietly about the wrong semester.\r\n             ",
        "responses": {
          "200": {
            "description": "The current term.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TermDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDto"
                }
              }
            }
          },
          "404": {
            "description": "No term in this school is flagged current.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/colleges": {
      "get": {
        "tags": [
          "Academic"
        ],
        "summary": "`GET /academic/colleges` — every college, by name.",
        "description": "Unfiltered, and paged like every other admin list — a college list really is a handful of rows\r\nper institution, but \"this table is small today\" is a property of the data rather than of the\r\nendpoint, and the two lists on this controller that were argued unbounded on exactly that\r\nreasoning (offerings, groups) are the two that grew per term. `code` is nullable: the\r\nroster source has no college code column, so the natural key is the normalized name and the\r\ncode exists only to be filled in later.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number, default 1. Out-of-range values are clamped, not refused.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 50, maximum 200; a larger value is clamped and the response says so.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of colleges, possibly empty.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/programs": {
      "get": {
        "tags": [
          "Academic"
        ],
        "summary": "`GET /academic/programs` — degree programmes, optionally narrowed to one college.",
        "parameters": [
          {
            "name": "collegeId",
            "in": "query",
            "description": "Omit to list them all. An id matching no college returns an empty list rather than everything —\r\na filter that silently stops filtering is how an \"invite this college\" flow ends up inviting\r\nthe institution.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number, default 1. Out-of-range values are clamped, not refused.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 50, maximum 200; a larger value is clamped and the response says so.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matching programmes, possibly empty.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgramDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgramDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgramDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/courses": {
      "get": {
        "tags": [
          "Academic"
        ],
        "summary": "`GET /academic/courses` — courses, optionally narrowed by college and search term.",
        "parameters": [
          {
            "name": "collegeId",
            "in": "query",
            "description": "Omit to list them all. A course's college is itself nullable, so an unattributed course matches\r\nno `collegeId` — which is correct: it is not known to be in one, rather than known not to\r\nbe.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Matched against <b>both</b> the code and the title. A user typing \"criminology\" is naming the\r\ntitle while one typing \"SSCI\" is naming the code, and asking which they meant is worse than\r\nsearching both. Substring, case-insensitive by the database's collation.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number, default 1. Out-of-range values are clamped, not refused.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 50, maximum 200; a larger value is clamped and the response says so.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matching courses, possibly empty.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/course-offerings": {
      "get": {
        "tags": [
          "Academic"
        ],
        "summary": "`GET /academic/course-offerings` — <b>the section grain, and the row an audience picker is\r\n             really looking for.</b> A course taught to two sections is two entries.",
        "description": "\r\n\r\n  <b>This is the endpoint that answers \"who is in BSCRIM 2-A\", and `students.section` is\r\n             not.</b> That column is the ADR-001 D-2 derived display cache: it is single-valued, and twelve\r\n             of the fifty-two students in the real roster sit in more than one section, so it is wrong for\r\n             roughly a quarter of them. An offering is the grain an invitation is actually issued at.\r\n             \r\n\r\n\r\n  `sectionKey` is published alongside `sectionName` because a <em>cohort</em> has no\r\n             row of its own — it is a key shared by every offering that cohort takes within a term, which is\r\n             why the group projection keys section groups on it. `sectionName` is null where the source\r\n             was blank (39 sample rows are); the key is a sentinel rather than null in that case, because a\r\n             nullable key component would have capped the table at one blank section under SQL Server's\r\n             unique index.\r\n             ",
        "parameters": [
          {
            "name": "termId",
            "in": "query",
            "description": "<b>Omit to get the current term, not every term.</b> Section names repeat each semester against\r\n             a different cohort, so an unscoped list stacked several distinct audiences under identical\r\n             names and grew with every import the institution had ever run — this endpoint returned every\r\n             offering of every term, which is the defect the default closes.\r\n            \r\n             \r\n<b>When no term is flagged current, the response is an empty page.</b> Zero current terms is a\r\n             real state — the flag is capped at one per school, not pinned at one, and between semesters\r\n             nobody has moved it yet. Widening back to every term in that case would restore the unbounded\r\n             read on the one day nobody is watching for it. Ask `GET /academic/terms/current` to tell\r\n             \"no offerings this term\" from \"no term is current\"; it 404s on the second.\r\n             ",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "courseId",
            "in": "query",
            "description": "Omit to list offerings of every course in the scoped term.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "section",
            "in": "query",
            "description": "The section's display name. <b>Normalized before it is compared</b> — `BSFS 2-A`,\r\n`bsfs2a` and `BSFS-2A` are one section — so the filter does not depend on how the\r\ncaller happened to type it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number, default 1. Out-of-range values are clamped, not refused.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 50, maximum 200; a larger value is clamped and the response says so.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matching offerings, possibly empty — including when no term is flagged current and\r\nnone was named.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOfferingDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOfferingDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOfferingDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TermDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TermDto"
            },
            "description": "This page's rows, in the endpoint's documented order. Empty is an ordinary answer — for an empty\r\nfilter and for a page past the end alike."
          },
          "page": {
            "type": "integer",
            "description": "The 1-based page actually served, which is <em>not</em> necessarily the one requested: an\r\nout-of-range `?page=` is clamped and this is where a client sees that it was.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "The page size actually applied, after the EAMS.Application.Dtos.Paging.MaxPageSize clamp. The field that\r\nmakes clamping discoverable rather than silent.",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "description": "Rows matching the filter across every page, counted in the database against the same filter that\r\nproduced Items. This is what a grid needs to render a page count; without it a\r\ncaller can only discover the end by walking into it.",
            "format": "int32"
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether another page exists after this one.\r\n            \r\n\r\n<b>Derived here rather than left to the client, because the client gets it wrong.</b>`page * pageSize < total` is right and `items.Count == pageSize` is the version\r\npeople write — which reports a further page whenever the last page happens to be exactly full,\r\nso a paging loop makes one extra request every time the total is a multiple of the page size.\r\nComputing it once, on the side that already holds the total, is a byte on the wire against a\r\nclass of off-by-one nobody finds in review.\r\n",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "One page of an admin list read, plus what the caller needs to ask for the next one.\r\n            \r\n\r\n<b>Every §6.2/§6.3 admin list returns this; `GET /attendance/live/{eventId}` deliberately does\r\nnot.</b> That endpoint pages by cursor (`since`/`cursor`/`hasMore`) because it is a\r\npolling delta rather than a list — a client asks \"what changed after this point\", not \"give me rows\r\n40 to 60\" — and its shape is frozen published contract under D-29/D-30/D-42. Offset paging over a\r\nlive feed also drops rows: an insert below the cursor shifts every subsequent row up by one and the\r\nnext page skips it. The two mechanisms coexisting is correct, not an inconsistency to tidy.\r\n"
      }
    },
    "securitySchemes": {
      "DeviceKey": {
        "type": "apiKey",
        "description": "The kiosk / capture credential, scoped to `attendance.capture` and nothing else.\n\nSend the **entire header value including the scheme word**:\n\n```\nAuthorization: DeviceKey eams_dk_<keyId>_<secret>\n```\n\nIssued once by `POST /api/v1/devices` and never retrievable afterwards — the server stores\nonly a hash. Store it in the platform keystore (`expo-secure-store` / Android Keystore),\nnever in `AsyncStorage`. Lose it and rotate with `POST /api/v1/devices/{id}/regenerate-key`.\n\nCase is significant and lower-case is the only accepted form for the token itself.\n\nFailures: `401 DeviceKeyMissing` / `DeviceKeyMalformed` / `DeviceKeyInvalid`,\n`403 DeviceKeyRevoked` / `DeviceInactive`, `429 RateLimited` (honour `Retry-After`).\n`DeviceKeyInvalid` deliberately does not distinguish an unknown key id from a wrong secret.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}